requirements.txt
anyio==4.4.0
apify_client==1.7.0
apify_shared==1.1.1
certifi==2024.6.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
sniffio==1.3.1
termcolor==2.5.0
typing_extensions==4.12.2
